+++ src/components/Timeline.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+
+export const Timeline = (): JSX.Element =>
+  <div>Timeline</div>;
+
+